% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRelMatrices.r
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Create heatmap from PLINK output}
\usage{
makeHeatmap(relMatName, graphTitle = "", fileExtension)
}
\arguments{
\item{relMatName}{Base name of the relationship matrix}

\item{graphTitle}{Title of the figure created as the output.}

\item{fileExtension}{the extension of the file created by PLINK. Should be set to "mdist" when using '--distance-matrix' and
"mibs" when using '--distance square ibs', or "genome" when using '--genome'}
}
\description{
Visualizes numerical matrices of n*n as heatmaps on the yellow to red scale.
Primarily intended to show IBD, IBS and distance matrices created by PLINK.
Also installs and loads the 'gplots' and 'RcolorBrewer' packages.
}
\examples{
makeHeatmap(outputName,graphTitle="Distance matrix",fileExtension="mdist")
}
\keyword{visualization}
