\name{mGenomics-package}
\alias{mGenomics-package}
\alias{mGenomics}
\docType{package}
\title{
mGenomics
}
\description{
These functions were used to build the metaGenomics layers into Synapse.
}
\details{
\tabular{ll}{
Package: \tab mGenomics\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-16\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Brig Mecham
Maintainer: Brig Mecham <brig.mecham@sagebase.org>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
