% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textScript.r
\name{myTestDogs}
\alias{myTestDogs}
\title{A Dog Function}
\usage{
myTestDogs(love = TRUE)
}
\arguments{
\item{love}{Do you love dogs? Defaults to TRUE.}
}
\description{
This function allows you to express your love of dogs.
Based on the Cat script of https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
}
\examples{
myTestDogs()
}
\keyword{dogs}
