% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcamds_plots.r
\name{pca_plots}
\alias{pca_plots}
\title{Plot PCA results}
\usage{
pca_plots(mds, classifier = 1, pca1, pca2, outname = "pcaPlot.png")
}
\arguments{
\item{mds}{Named object of the eigenvectors produced by the get() function}

\item{classifier}{Use uncertain, default value = 1}

\item{pca1}{The eigenvector and its explained variance plotted on the X axis}

\item{pca2}{The eigenvector and its explained variance plotted on the Y axis}

\item{outname}{The name of the png file to be saved to the computer. Default value is "pcaPlot.png"}
}
\description{
Saves a picture of a PCA plot to the HDD with various formatting adjustments.
The original script supplied by Solomon A. Boison
}
\examples{
pca_plots(eigen_vec="eigenvec_populations",pca1=c(1,paste(eigen_values[1],"\%")),
pca2=c(2,paste(eigen_values[2],"\%")))
}
\keyword{PCA}
