% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped2bped.r
\name{ped2bped}
\alias{ped2bped}
\title{PLINK file format change from ped to binary-ped}
\usage{
ped2bped(dataIn, otherParameters = "--dog --allow-no-sex --nonfounders")
}
\arguments{
\item{dataIn}{Input data ped and map file name as required by PLINK's --file parameter}

\item{otherParameters}{Optional parameters that are necessary for a specific analysis. 
Default value set to " --dog --allow-no-sex --nonfounders"}
}
\description{
Creates and addtional binary ped (.bed, .bim and .fam) files using the - -make-bed parameter of PLINK. 
The name of the output file is the same as the name of the input file.
}
\examples{
ped2bped("myInputFileName")
}
\keyword{plink}
