% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkQC.r
\name{plinkQC}
\alias{plinkQC}
\title{Quality control with PLINK}
\usage{
plinkQC(dataIn)
}
\arguments{
\item{dataIn}{Input data file name as required by PLINK's --file parameter}
}
\description{
This function is to do a quality control based on user specified paramters.
A version of plink.exe compiled for Windows is already supplied with the R package.
Creates both ped and binary ped outputs of the Quality controlled data. 
WARNING: In development: allows to specify the data name only, binary ped files expected
}
\examples{
plinkQC("myInputFileName")
}
\keyword{plink}
