% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROH.r
\name{plotROH}
\alias{plotROH}
\title{Visualize ROH results}
\usage{
plotROH(dataIn)
}
\arguments{
\item{dataIn}{The name of the R data frame created by runROH(), containing the inbreeding result for each indvidual.}
}
\description{
Visualizes the results of a ROH analysis made with the runROH() function
}
\examples{
plotROH("afterRohAnalysis")
}
\keyword{ROH}
