% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCA.r
\name{runPCA}
\alias{runPCA}
\title{Compute and plot principal components}
\usage{
runPCA(dataIn, useK = 4, eigVecOnX = 1, eigVecOnY = 2,
  outname = "pcaPlot.png")
}
\arguments{
\item{dataIn}{Name of the input data file. The function is tested with PLINK's .mdist files.
The input should be the full filename e.g. "myResults.mdist"}

\item{useK}{the maximum dimension of the space which the data are to be represented in
Must be in {1, 2, ., n-1}. The default value is 4}

\item{eigVecOnX}{The eigenvector to be plotted on X axis. Default is 1}

\item{eigVecOnY}{The eigenvector to be plotted on X axis. Default is 2}

\item{outname}{The name of the png file to be saved to the computer. Default value is "pcaPlot.png"}
}
\description{
Computes the principal components using the cmdscale function and plots the results 
using the pca_plots function of the MGenomics package
}
\examples{
runPCA("myResults.mdist")
}
\keyword{plink}
