% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlink.r
\name{runPlink}
\alias{runPlink}
\title{Run PLINK}
\usage{
runPlink(plinkParameters)
}
\arguments{
\item{plinkParameters}{Parameters specified according to the PLINK manual}
}
\description{
Executes the PLINK program supplied with the package
}
\examples{
runPlink("--bfile inData --recode --out outData")
}
\keyword{plink}
