% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROH.r
\name{runROH}
\alias{runROH}
\title{ROH analysis}
\usage{
runROH(dataIn, outName, lParameter, minRohLength, maxGap, maxMissing,
  maxHetero, genomeLength = 2201660000)
}
\arguments{
\item{dataIn}{Input data ped and map file name fro ROH analyses}

\item{outName}{Base name for output files.}

\item{lParameter}{ROH length according to the cgaTOH manual.}

\item{minRohLength}{Minimum length of ROH segment in base pairs}

\item{maxGap}{Maximum physcal gap alloved between adjascent SNPs in ROH in base pairs}

\item{maxMissing}{Maximum number of missing SNPs allowed in a ROH run}

\item{maxHetero}{Maximum number of heterozygous SNPs allowed in a ROH run}

\item{genomeLength}{The length of the genome covered by SNPs in base pairs 
that should be considered for calcualtion of the genomic inbreeding. 
The default value is the autosomal length of the dog autosome, 2201660000 base pairs.}
}
\description{
Executes a ROH analysis using the cgaTOH program
}
\examples{
runROH("afterQC",outName = "test", minRohLength = 1000000, lParameter = 15, maxGap = 10000, maxMissing = 0, maxHetero = 0)
}
\keyword{ROH}
